Rename links controller to more general shares controller
